Add kubernetes template files.
This commit introduces a set of Kubernetes template files that would
be used initially to handle variables pertaining to avi controller
userid, password, controller version, controller ip etc.

The configmap is used for non-sensitive data and is kept as a flat
structure for now. Once we evolve this architecture, we would want
to move to a CRD for the config options.