Support for vsvip
This PR adds support for VSVIP as a separate object.
We don't add the vip related info inside the vs object itself, instead
use a separate vsvip object.

When a new ingress is synced, the vsvip object is updated with the
new fqdn list. Also if a ingress is deleted, the fqdn list is
updated.

This required us to track the ingress to host mapping to handle
the delete scenarios.