AV-82537: Updates during AKO shutdown.
This PR takes care of reconciliation logic while AKO is shutdown.
On every bootup of AKO, we fetch the objects from the controller
and cache them. We run a fullsync on every bootup. Previously
we built the objects in Layer 2 and synced them directly to Layer 3 as a
part of the full sync.

As a part of this commit, we reverse the process. We let the models
get built in Layer 2 and then we publish all the models present in the
cache for a re-sync followed by the model objects which got generated.

This ensures that if there are stale objects while AKO was shutdown,
those objects are deleted from the Avi controller.