AV-85888: Cache robustness
This fix makes 2 changes to the existing cache model for SNIs:

 - During SNI relationship building if due to some reason we
do not find a HTTP Policy reference in the indvidual httppol cache,
we do an explicit GET on the httpcache once per VS page.
Same is repeated for PG as well, on a per PG level.

 - During deletion of the SNI VS if we detect a flaw in the SNI
cache, we do a manual refresh to get the latest cache.