AV-88707: Test cases for routes with edge termination (#213)
* AV-88707: Test cases for routes with edge termination

Other than test cases, there are following chnages:

- Added type for CA cert and sslkeycert

- Fixed a potential bug, if svc for a secure ingress/route gets processed before
processing the route/ingress. This would be rare as we would face the issue
only if route/ingress and svc gets created and deleted very fast. The secret
mapping may not get updated and after deletion of route/ingress we may
see stale SNI. Saw this only in unit test and not in actual route/svc
create/delete operations.

To solve this - updating secret mapping during hostpath processing of
the route/ingress.

* Added test case for multiple secure routes in different namesapces