Stop saving model from layer2 if disable sync is set to true
In case of repeated enable and disable sync in a cluster with large number of objects,
layer2 may be processing an object after sync has been disabled from layer1.
In this case layer1 would set the model to nil, but layer2 can save and overrite the value.

Later when layer3 picks up the key to process, it won't find nil model and won't delete
all avi objects.

Checking if disableSync is set to before saving model from layer2 to handle this.