Re-org helm values
This commit re-organizes the values.yaml file into distinct buckets
for each type of settings.

We would like to start AKO with minimal edits to the values.yaml.

Currently the following values are mandatory:

 - ClusterName.
 - ControllerIP
 - Controller username
 - Controlller password
 - Placement network details.
 - NetworkName/Subnet details

In the subsequent PRs, we might want to make the placement network
details and the networkName/subnet details optional.