Refactor vipNetworkList to include CIDR
This updates the VipNetworkList schemas in helm's values.yaml,
AviInfraSetting CRD and ako-operator's AKOConfig, which now includes
an extra `cidr` field that allows users to choose specific subnet in
a network. This replaces the previous subnetIP/subnetPrefix fields.
The AviInfraSetting and values.yaml route has validations for the CIDR
string to be of a certain format.
This also removes unused EastWest fields across AKO.