VIP per namespace for EVH in VCF
In VCF, we would create VIP per namespace, adding support for that in this PR.
We would look for an environment variable USE_VCF to enable this feature.

In this case, a mapping between namespace and hostname is saved to find out if
a hostname is present in multiple namespaces.

The parent VS name is modified in this case to include namespace instead of shard.