Status update fixes
This commit simplifies switching mechanisms for status updates
based on the serviceMetadata information, and helps identify
corresponding to which object the status update is bound for.
Fixes include - not allowing status update for SNI VS pools,
correct status update logic in case of vsvip updates.
For VSVIP related status updates, it needs to happen only in case
of VSVIP PUT calls.
Fixes: AV-128633

.