Update annotation in AKO statefulset after avi object deletion through deleteConfig (#689)
Currently status of ako statefulset is updated with condition values after objects are deleted
by setting deleteConfig to true. However, in kubernetes 1.22 this can not be done because kubernetes
automatically removes this status. To get around this problem annotation of the statefulset would be
updated instead of status.

When object deletion starts we would set the annotation as:
  annotations:
    AviObjectDeletionStatus: Started

After object deletion is completed, this would be changed to
    AviObjectDeletionStatus: Done

(cherry picked from commit 71002106456eaedbd3bd73fb5ead0593e76985a5)