Remove overwriting controllerVersion on POST/PUT/DELETE api calls
This commit removes the controller version that is being overwritten
on individual API calls to the controller instead of using the api
version which was used to initialise the session to the controller.
With these changes, all REST layer API calls end up using the version
which was used while initializing the session.