Fix Service LB handling with Gateways
This commit fixes issues with using Gateways and ServiceLB at the same time.
The bug lead to improper dedicated VS creation of Service of type LB.
This also fixes an issue with L4 VSVIP when autoFqdn is disabled and no
subDomains are configured in the dnsProfile, which was leading to an empty
string fqdn programming.